Refactor Dockerfile and main.py for improved clarity; update sample.env and settings.py for consistency; add requests.py for API interactions.
