+++ sample.env
@@ -1,2 +1,2 @@
-TOKEN=INSERT_TOKEN_HERE
-API_URL=localhost:8000
+DISCORD_BOT_TOKEN=discord_token
+API_URL=host_address
