+++ telibot/core/__init__.py
@@ -1,5 +1,5 @@
 """This module contains the core functionalities of the bot."""
 
-from . import activies, client, utils
+from . import activies, client, requests
 
-__all__ = ["activies", "client", "utils"]
+__all__ = ["activies", "client", "requests"]
